pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "birdcam"
version = "0.1.0"
description = "Bird detection camera system for Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",           # for FastAPI test client
]
pi = [
    "picamera2>=0.3.19",
]

[project.scripts]
birdcam = "birdcam.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
